table statements for the ScaledUp version
This will totally replace the back end tables once CounterMeme gets big. The idea is reduced searching complexity and php calls. Basically to find for instance all the memes created by a user, the php in the ScaledUp version will read "foreach meme in memestash, echo meme..." while the ScaledDown php would require searching through every meme that has ever been created on the site to display a user's memestash. 
Also of note, the $username_ is made to mimic the php call that will be used as the table names will begin with the new user's username so we are sure each like table has a unique identifier.